Create a calculator class under the main package
